get build working on CircleCI and deploying in PWS
We've removed heroku-postbuild as it was always failing as it runs when
the app stages and has no database credentials available.

We also had to move several packages from prod to dev as the app was
exhausting the available disk space when deployed to PWS.  The
dependencies section should really only be things we need to run the API
app - testing and compiling assets should be taken care of by the CI
build.

We've also added the CI status badge to readme! Yaaay!

[#141636805]